@@ -25,12 +25,7 @@
 
         public virtual IQueryable<TEntity> AllAsNoTracking() => this.DbSet.AsNoTracking();
 
-        public virtual Task<TEntity> GetByIdAsync(params object[] id) => this.DbSet.FindAsync(id);
-
-        public virtual void Add(TEntity entity)
-        {
-            this.DbSet.Add(entity);
-        }
+        public virtual Task AddAsync(TEntity entity) => this.DbSet.AddAsync(entity).AsTask();
 
         public virtual void Update(TEntity entity)
         {
@@ -43,13 +38,22 @@
             entry.State = EntityState.Modified;
         }
 
-        public virtual void Delete(TEntity entity)
-        {
-            this.DbSet.Remove(entity);
-        }
+        public virtual void Delete(TEntity entity) => this.DbSet.Remove(entity);
 
         public Task<int> SaveChangesAsync() => this.Context.SaveChangesAsync();
 
-        public void Dispose() => this.Context.Dispose();
+        public void Dispose()
+        {
+            this.Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+            if (disposing)
+            {
+                this.Context?.Dispose();
+            }
+        }
     }
 }
