@@ -1,17 +1,15 @@
 ﻿namespace LoveAdvisor.Data.Seeding
 {
     using System;
-    using System.Linq;
+    using System.Collections.Generic;
+    using System.Threading.Tasks;
 
-    using LoveAdvisor.Common;
-    using LoveAdvisor.Data.Models;
-
-    using Microsoft.AspNetCore.Identity;
     using Microsoft.Extensions.DependencyInjection;
+    using Microsoft.Extensions.Logging;
 
-    public static class ApplicationDbContextSeeder
+    public class ApplicationDbContextSeeder : ISeeder
     {
-        public static void Seed(ApplicationDbContext dbContext, IServiceProvider serviceProvider)
+        public async Task SeedAsync(ApplicationDbContext dbContext, IServiceProvider serviceProvider)
         {
             if (dbContext == null)
             {
@@ -23,41 +21,19 @@
                 throw new ArgumentNullException(nameof(serviceProvider));
             }
 
-            var roleManager = serviceProvider.GetRequiredService<RoleManager<ApplicationRole>>();
-            Seed(dbContext, roleManager);
-        }
+            var logger = serviceProvider.GetService<ILoggerFactory>().CreateLogger(typeof(ApplicationDbContextSeeder));
 
-        public static void Seed(ApplicationDbContext dbContext, RoleManager<ApplicationRole> roleManager)
-        {
-            if (dbContext == null)
+            var seeders = new List<ISeeder>
+                          {
+                              new RolesSeeder(),
+                              new SettingsSeeder(),
+                          };
+
+            foreach (var seeder in seeders)
             {
-                throw new ArgumentNullException(nameof(dbContext));
-            }
-
-            if (roleManager == null)
-            {
-                throw new ArgumentNullException(nameof(roleManager));
-            }
-
-            SeedRoles(roleManager);
-        }
-
-        private static void SeedRoles(RoleManager<ApplicationRole> roleManager)
-        {
-            SeedRole(GlobalConstants.AdministratorRoleName, roleManager);
-        }
-
-        private static void SeedRole(string roleName, RoleManager<ApplicationRole> roleManager)
-        {
-            var role = roleManager.FindByNameAsync(roleName).GetAwaiter().GetResult();
-            if (role == null)
-            {
-                var result = roleManager.CreateAsync(new ApplicationRole(roleName)).GetAwaiter().GetResult();
-
-                if (!result.Succeeded)
-                {
-                    throw new Exception(string.Join(Environment.NewLine, result.Errors.Select(e => e.Description)));
-                }
+                await seeder.SeedAsync(dbContext, serviceProvider);
+                await dbContext.SaveChangesAsync();
+                logger.LogInformation($"Seeder {seeder.GetType().Name} done.");
             }
         }
     }
