@@ -1,21 +1,18 @@
 ﻿namespace LoveAdvisor.Services.Messaging
 {
+    using System.Collections.Generic;
     using System.Threading.Tasks;
 
-    // This class is used by the application to send Email and SMS
-    // when you turn on two-factor authentication in ASP.NET Identity.
-    // For more details see this link https://go.microsoft.com/fwlink/?LinkID=532713
-    public class NullMessageSender : IEmailSender, ISmsSender
+    public class NullMessageSender : IEmailSender
     {
-        public Task SendEmailAsync(string email, string subject, string message)
+        public Task SendEmailAsync(
+            string from,
+            string fromName,
+            string to,
+            string subject,
+            string htmlContent,
+            IEnumerable<EmailAttachment> attachments = null)
         {
-            // Plug in your email service here to send an email.
-            return Task.CompletedTask;
-        }
-
-        public Task SendSmsAsync(string number, string message)
-        {
-            // Plug in your SMS service here to send a text message.
             return Task.CompletedTask;
         }
     }
