@@ -1,11 +1,8 @@
 ﻿namespace LoveAdvisor.Web.Controllers
 {
-    using Microsoft.AspNetCore.Authentication.JwtBearer;
-    using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.Mvc;
 
-    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
-    public abstract class BaseController : Controller
+    public class BaseController : Controller
     {
     }
 }
